FPII-2291 : Remote code execution vulnerability in MediaMuxer CVE-2016-3863 A-29161888
The convertMessageToMetaData function in Utils.cpp of libstagefright
assumes that the length of avcc will be less than 1024. However, there is no limit to the length, which could lead to a stack overflow in the calling application.

The fix is designed to ensure the buffer is large enough to handle the
length of avcc, and therefore prevent the stack overflow.

Change-Id: I85a5b12d6586b9b7b355ed5a7f5cfa8bcf3dabcd