FPII-2479: Update: Denial of service vulnerability in MediaServer CVE-2016-3754 A-28615448
Within a specially crafted mpeg4 file, the psshsize can be set to a value that is larger than can be allocated,
which can result in a mediaserver crash. This provides an additional layer of protection on top of the original fix for the issue.

The updated fix is designed to limit mp4 atom size to a reasonable value.

Change-Id: I00a2c377825f8a6bd4194ef53b1e4fee103fb5e5