FPII-2465 Elevation of privilege vulnerability in Mediaserver CVE-2016-6705 A-30907212
Obtaining a reference pointer, instead of a strong pointer, to the service interface could lead to a race condition
where the service pointer could be cleared by another thread while in use. This could cause a use-after-free vulnerability.

The fix is designed to get a strong pointer to the service interface.

Change-Id: I6943b33c697a3f078e06d588ac934dbad892d377